Update README.md
closes #81
The remaining columns HR_Avg_1, HR_Incr_1 ... 3.  Create a piecewise linear heat rate curve by defining the average heat rate at the dispatch level defined by `Output_pct_0 * PMax MW` and the slope of the segment between dispatch points 0 and 1 as HR_Incr_1, the slope of the segment between points 1 and 2 as HR_Incr_2, and the slope of the segment between points 2 and 3 as HR_Incr_3.